+++ CHANGELOG.md
@@ -23,6 +23,7 @@ https://github.com/nwnxee/unified/compare/build8193.37.13...HEAD
 - Player: GetOpenStore()
 - Creature: GetNumberOfBonusSpells(), ModifyNumberBonusSpells()
 - Utils: RawPrint()
+- Utils: SetStartingLocation()
 
 ### Changed
 - Damage: Added bRangedAttack to the NWNX_Damage_AttackEventData struct.